reorganize multicluster tests
